ai_module/neural_network-main/requirements.txt
numpy>=1.21.0
streamlit>=1.28.0
streamlit-drawable-canvas>=0.9.0
opencv-python>=4.5.0
matplotlib>=3.5.0
pillow>=8.3.0
jupyter>=1.0.0 